Adjust loading and saving
Use ServerStartingEvent to create and/or load NetworkData. Add networks to server cache in NetworkData#readFromNBT as opposed to loading from NBT, caching to a list, loading into the server cache, then clearing the list. Also, since the overworld's perWorldStorage is the same as its mapStorage, handling the loading of 'old' saving mechanisms shouldn't be necessary.